requirements.txt
gitpython
pydantic
mcp
